Move get_object(s) to XoProcedures
