+++ src/api/mod.rs
@@ -3,26 +3,16 @@ pub mod token;
 pub mod vm;
 pub mod xo;
 
-use std::{collections::BTreeMap, sync::Arc};
+use std::sync::Arc;
 
-use jsonrpsee_types::{traits::SubscriptionClient, DeserializeOwned, JsonValue, Subscription};
+use jsonrpsee_types::{traits::SubscriptionClient, DeserializeOwned, Subscription};
 use jsonrpsee_ws_client::{WsClient, WsClientBuilder};
 
-use crate::{
-    procedure_object,
-    types::{XoObject, XoObjectMap},
-    RpcError,
-};
+use crate::RpcError;
 
 use self::{
     session::SessionProcedures, token::TokenProcedures, vm::VmProcedures, xo::XoProcedures,
 };
-
-#[derive(Debug)]
-pub enum GetSingleObjectError {
-    MultipleMatches,
-    Rpc(RpcError),
-}
 
 /// Client used to communicate with Xen Orchestra's API
 ///
@@ -55,7 +45,7 @@ pub enum GetSingleObjectError {
 ///         .expect("Failed to sign in");
 ///
 ///     let all_vms: BTreeMap<VmId, Vm<OtherInfo>> =
-///         con.get_objects(None, None).await.expect("Failed to list VMs");
+///         con.xo.get_objects(None, None).await.expect("Failed to list VMs");
 ///
 ///     let test_vms = all_vms
 ///         .iter()
@@ -120,49 +110,5 @@ impl Client {
         T: DeserializeOwned,
     {
         self.inner.subscribe_to_method::<T>("all").await
-    }
-
-    /// Get all objects of specified type from server
-    /// * `R` is a type that can represent that collection of objects
-    /// * `filter` is an optional filter
-    /// * `limit` is an optional max limit on number of results
-    pub async fn get_objects<R: XoObjectMap>(
-        &self,
-        filter: impl Into<Option<serde_json::Map<String, JsonValue>>>,
-        limit: impl Into<Option<usize>>,
-    ) -> Result<R, RpcError> {
-        let mut filter = filter.into().unwrap_or_default();
-        filter.insert("type".to_string(), R::Object::OBJECT_TYPE.into());
-
-        self.xo.get_all_objects(filter, limit).await
-    }
-
-    /// Get single object of specified type from server
-    /// * `R` is a type that can represent that type of object
-    /// * `id` is the id of the object
-    pub async fn get_object<R: XoObject>(
-        &self,
-        id: R::IdType,
-    ) -> Result<Option<R>, GetSingleObjectError>
-    where
-        R::IdType: Ord,
-    {
-        let filter = procedure_object!(
-            "id" => id.clone(),
-            "type" => R::OBJECT_TYPE.to_string()
-        );
-
-        // TODO: Can we get rid of the BTreeMap here?
-        let mut result: BTreeMap<R::IdType, R> = self
-            .xo
-            .get_all_objects(filter, Some(2))
-            .await
-            .map_err(GetSingleObjectError::Rpc)?;
-
-        match result.remove(&id) {
-            None => Ok(None),
-            Some(vm) if result.is_empty() => Ok(Some(vm)),
-            _ => Err(GetSingleObjectError::MultipleMatches),
-        }
     }
 }
